# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TransitOps Smart Transport Operations Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all trips
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TripStatus"
        - name: vehicleId
          in: query
          schema:
            type: integer
        - name: driverId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip (Draft state)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Trip created in Draft state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "400":
          description: Validation error (capacity exceeded, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a trip by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update a trip (only in Draft state)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}/dispatch:
    post:
      operationId: dispatchTrip
      tags: [trips]
      summary: Dispatch a trip (Draft → Dispatched, sets vehicle + driver On Trip)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dispatched trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "400":
          description: Business rule violation (capacity, availability, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{id}/complete:
    post:
      operationId: completeTrip
      tags: [trips]
      summary: Complete a trip (Dispatched → Completed, restores vehicle + driver)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripCompletion"
      responses:
        "200":
          description: Completed trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{id}/cancel:
    post:
      operationId: cancelTrip
      tags: [trips]
      summary: Cancel a trip (restores vehicle + driver if dispatched)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"

  # ─── MAINTENANCE ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string

    # ─── ENUMS ──────────────────────────────────────────────────────────────────
...
    TripStatus:
      type: string
      enum: [draft, dispatched, completed, cancelled]
...
    Trip:
      type: object
      required: [id, sourceLocation, destinationLocation, cargoWeightKg, plannedDistanceKm, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        sourceLocation:
          type: string
        destinationLocation:
          type: string
        vehicleId:
          type: ["integer", "null"]
        driverId:
          type: ["integer", "null"]
        vehicleName:
          type: ["string", "null"]
        driverName:
          type: ["string", "null"]
        registrationNumber:
          type: ["string", "null"]
        cargoWeightKg:
          type: number
        plannedDistanceKm:
          type: number
        actualDistanceKm:
          type: ["number", "null"]
        plannedStartAt:
          type: ["string", "null"]
          format: date-time
        dispatchedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        cancelledAt:
          type: ["string", "null"]
          format: date-time
        finalOdometerKm:
          type: ["number", "null"]
        revenue:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/TripStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TripInput:
      type: object
      required: [sourceLocation, destinationLocation, vehicleId, driverId, cargoWeightKg, plannedDistanceKm]
      properties:
        sourceLocation:
          type: string
          minLength: 1
        destinationLocation:
          type: string
          minLength: 1
        vehicleId:
          type: integer
        driverId:
          type: integer
        cargoWeightKg:
          type: number
          minimum: 0
        plannedDistanceKm:
          type: number
          minimum: 0.1
        plannedStartAt:
          type: string
          format: date-time
        revenue:
          type: number
          minimum: 0
        notes:
          type: string
    TripUpdate:
      type: object
      properties:
        sourceLocation:
          type: string
        destinationLocation:
          type: string
        vehicleId:
          type: integer
        driverId:
          type: integer
        cargoWeightKg:
          type: number
          minimum: 0
        plannedDistanceKm:
          type: number
          minimum: 0.1
        plannedStartAt:
          type: string
          format: date-time
        revenue:
          type: number
          minimum: 0
        notes:
          type: string
    TripCompletion:
      type: object
      required: [actualDistanceKm, finalOdometerKm]
      properties:
        actualDistanceKm:
          type: number
          minimum: 0
        finalOdometerKm:
          type: number
          minimum: 0
        revenue:
          type: number
          minimum: 0
        notes:
          type: string

    # ─── MAINTENANCE ────────────────────────────────────────────────────────────
...