# lib/api-zod/src/generated/types/trip.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TransitOps Smart Transport Operations Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TripStatus } from './tripStatus';

export interface Trip {
  id: number;
  sourceLocation: string;
  destinationLocation: string;
  /** @nullable */
  vehicleId?: number | null;
  /** @nullable */
  driverId?: number | null;
  /** @nullable */
  vehicleName?: string | null;
  /** @nullable */
  driverName?: string | null;
  /** @nullable */
  registrationNumber?: string | null;
  cargoWeightKg: number;
  plannedDistanceKm: number;
  /** @nullable */
  actualDistanceKm?: number | null;
  /** @nullable */
  plannedStartAt?: Date | null;
  /** @nullable */
  dispatchedAt?: Date | null;
  /** @nullable */
  completedAt?: Date | null;
  /** @nullable */
  cancelledAt?: Date | null;
  /** @nullable */
  finalOdometerKm?: number | null;
  /** @nullable */
  revenue?: number | null;
  /** @nullable */
  notes?: string | null;
  status: TripStatus;
  createdAt: Date;
  updatedAt: Date;
}